consolidate & delete `types` files
