--- src/components/command-bar/components/divider/types.ts
@@ -1,10 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-interface CommandBarDividerProps {
-	className?: string
-}
-
-export type { CommandBarDividerProps }
